requirements.txt
asgiref==3.4.1
certifi==2021.10.8
charset-normalizer==2.0.9
defusedxml==0.7.1
Django==3.2.10
django-environ==0.8.1
djangorestframework==3.13.0
djangorestframework-link-header-pagination==0.1.1
httpie==2.6.0
idna==3.3
Pygments==2.10.0
PySocks==1.7.1
pytz==2021.3
requests==2.26.0
requests-toolbelt==0.9.1
sqlparse==0.4.2
typing-extensions==4.0.1
urllib3==1.26.7
